chore(borrow-navigation): adding typing within common libs
