+++ apps/ledger-live-mobile/src/mvvm/features/Borrow/screens/BorrowLiveApp/BorrowLiveAppWrapper.tsx
@@ -1,13 +1,9 @@
 import { WalletAPICustomHandlers } from "@ledgerhq/live-common/wallet-api/types";
-import {
-  BorrowSwapNavigationParams,
-  createBorrowNavigateHandler,
-} from "@ledgerhq/live-common/wallet-api/Borrow/navigate";
+import { createBorrowNavigateHandler } from "@ledgerhq/live-common/wallet-api/Borrow/navigate";
+import type { BorrowSwapNavigationParams } from "@ledgerhq/live-common/wallet-api/Borrow/types";
 import React, { useEffect, useMemo } from "react";
 import { BorrowLiveAppView } from ".";
 import { useBorrowLiveAppViewModel } from "LLM/features/Borrow/screens/BorrowLiveApp/useBorrowLiveAppViewModel";
-
-export type { BorrowSwapNavigationParams };
 
 type BorrowLiveAppWrapperProps = Readonly<{
   action?: "go-back";
