+++ libs/ledger-live-common/.unimportedrc.json
@@ -306,6 +306,7 @@
     "src/wallet-api/constants.ts",
     "src/wallet-api/converters.ts",
     "src/wallet-api/Borrow/navigate.ts",
+    "src/wallet-api/Borrow/types.ts",
     "src/wallet-api/CustomLogger/client.ts",
     "src/wallet-api/CustomLogger/server.ts",
     "src/wallet-api/CustomDeeplink/server.ts",
